Remove irrelevant files
