+++ treeple/stats/__init__.py
@@ -1,14 +1,8 @@
-from .forestht import (
-    build_coleman_forest,
-    build_cv_forest,
-    build_oob_forest,
-    build_permutation_forest,
-)
-from .monte_carlo import PermutationTest
+from .baseline import build_cv_forest, build_permutation_forest
+from .forest import build_coleman_forest, build_oob_forest
 from .permuteforest import PermutationHonestForestClassifier
 
 __all__ = [
-    "PermutationTest",
     "build_cv_forest",
     "build_oob_forest",
     "build_coleman_forest",
