+++ treeple/stats/tests/meson.build
@@ -1,6 +1,7 @@
 python_sources = [
   '__init__.py',
-  'test_forestht.py',
+  'test_forest.py',
+  'test_baseline.py',
   'test_coleman.py',
   'test_utils.py',
   'test_permuteforest.py',
