+++ treeple/stats/utils.py
@@ -15,9 +15,9 @@ from sklearn.metrics import (
     roc_auc_score,
     roc_curve,
 )
-from sklearn.utils.validation import check_is_fitted, check_X_y
+from sklearn.utils.validation import check_is_fitted
 
-from treeple._lib.sklearn.ensemble._forest import BaseForest, ForestClassifier
+from treeple._lib.sklearn.ensemble._forest import BaseForest
 
 BOTTLENECK_AVAILABLE = False
 if "bottleneck" in sys.modules:
@@ -157,66 +157,6 @@ def _non_nan_samples(posterior_arr: ArrayLike) -> ArrayLike:
     # Invert the boolean mask to get indices without NaN values
     nonnan_indices = np.where(~nan_indices)[0]
     return nonnan_indices
-
-
-def _compute_null_distribution_perm(
-    X_train: ArrayLike,
-    y_train: ArrayLike,
-    X_test: ArrayLike,
-    y_test: ArrayLike,
-    covariate_index: ArrayLike,
-    est: ForestClassifier,
-    metric: str = "mse",
-    n_repeats: int = 1000,
-    seed: Optional[int] = None,
-    **metric_kwargs,
-) -> ArrayLike:
-    """Compute null distribution using permutation method.
-
-    Parameters
-    ----------
-    X_test : ArrayLike of shape (n_samples, n_features)
-        The data matrix.
-    y_test : ArrayLike of shape (n_samples, n_outputs)
-        The output matrix.
-    covariate_index : ArrayLike of shape (n_covariates,)
-        The indices of the covariates to permute.
-    est : ForestClassifier
-        The forest that will be used to recompute the test statistic on the permuted data.
-    metric : str, optional
-        The metric, which to compute the null distribution of statistics, by default 'mse'.
-    n_repeats : int, optional
-        The number of times to sample the null, by default 1000.
-    seed : int, optional
-        Random seed, by default None.
-    """
-    rng = np.random.default_rng(seed)
-    X_test, y_test = check_X_y(X_test, y_test, ensure_2d=True, multi_output=True)
-    # n_samples_test = len(y_test)
-    n_samples_train = len(y_train)
-    metric_func = METRIC_FUNCTIONS[metric]
-
-    # pre-allocate memory for the index array
-    train_index_arr = np.arange(n_samples_train, dtype=int).reshape(-1, 1)
-    # test_index_arr = np.arange(n_samples_test, dtype=int).reshape(-1, 1)
-
-    null_metrics = np.zeros((n_repeats,))
-
-    for idx in range(n_repeats):
-        rng.shuffle(train_index_arr)
-        perm_X_cov = X_train[train_index_arr, covariate_index]
-        X_train[:, covariate_index] = perm_X_cov
-
-        # train a new forest on the permuted data
-        # XXX: should there be a train/test split here? even w/ honest forests?
-        est.fit(X_train, y_train.ravel())
-        y_pred = est.predict(X_test)
-
-        # compute two instances of the metric from the sampled trees
-        metric_val = metric_func(y_test, y_pred, **metric_kwargs)
-
-        null_metrics[idx] = metric_val
-    return null_metrics
 
 
 def _compute_null_distribution_coleman(
